backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
sqlalchemy
aiomysql
motor
python-dateutil
pytz
python-jose
httpx
openai
numpy
pandas
torch
